requirements.txt
ttkbootstrap>=1.10.1
python-docx>=0.8.11
PyPDF2>=3.0.0
requests>=2.31.0
nltk>=3.8
textblob>=0.17.0
Pillow>=10.0.0
imagehash>=4.3.2
send2trash>=1.8.0
